fix blur issue
