--- src/context/ThemeContext.tsx
@@ -1,41 +1,0 @@
-import { createContext, useState, useEffect, useMemo } from "react";
-
-interface ThemeContextType {
-  theme: string;
-  toggleTheme: () => void;
-}
-
-export const ThemeContext = createContext<ThemeContextType>({
-  theme: "dark",
-  toggleTheme: () => {},
-});
-
-export const ThemeProvider: React.FC<{ children: React.ReactNode }> = ({
-  children,
-}) => {
-  const [theme, setTheme] = useState<string>(() => {
-    const storedTheme = localStorage.getItem("theme");
-    return storedTheme ? storedTheme : "dark";
-  });
-
-  const toggleTheme = () => {
-    setTheme((theme) => (theme === "light" ? "dark" : "light"));
-  };
-
-  useEffect(() => {
-    localStorage.setItem("theme", theme);
-  }, [theme]);
-
-  const themeContext: ThemeContextType = useMemo(() => {
-    return {
-      theme,
-      toggleTheme,
-    };
-  }, [theme, toggleTheme]);
-
-  return (
-    <ThemeContext.Provider value={themeContext}>
-      {children}
-    </ThemeContext.Provider>
-  );
-};
