+++ src/sections/Home/Home.css
@@ -4,7 +4,6 @@
   position: relative;
   row-gap: 4.5rem;
   padding-top: 2rem;
-  z-index: 10;
 }
 
 .home__data {
